Inline shell layout into main window
