--- Veriado.WinUI/Views/MainShell.xaml
@@ -1,109 +1,0 @@
-<UserControl
-    x:Class="Veriado.WinUI.Views.MainShell"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:conv="using:Veriado.WinUI.Converters"
-    xmlns:views="using:Veriado.WinUI.Views"
-    xmlns:winui="using:Microsoft.UI.Xaml.Controls">
-
-    <UserControl.Resources>
-        <conv:BooleanToVisibilityConverterEx x:Key="BoolToVisibilityEx" />
-    </UserControl.Resources>
-
-    <Grid x:Name="LayoutRoot">
-        <VisualStateManager.VisualStateGroups>
-            <VisualStateGroup>
-                <VisualState x:Name="Narrow">
-                    <VisualState.StateTriggers>
-                        <AdaptiveTrigger MinWindowWidth="0" />
-                    </VisualState.StateTriggers>
-                    <VisualState.Setters>
-                        <Setter Target="PreviewDetailPane.Visibility" Value="Collapsed" />
-                    </VisualState.Setters>
-                </VisualState>
-                <VisualState x:Name="Wide">
-                    <VisualState.StateTriggers>
-                        <AdaptiveTrigger MinWindowWidth="1200" />
-                    </VisualState.StateTriggers>
-                    <VisualState.Setters>
-                        <Setter Target="PreviewDetailPane.Visibility" Value="Visible" />
-                    </VisualState.Setters>
-                </VisualState>
-            </VisualStateGroup>
-        </VisualStateManager.VisualStateGroups>
-        <Grid.RowDefinitions>
-            <RowDefinition Height="*" />
-            <RowDefinition Height="Auto" />
-        </Grid.RowDefinitions>
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="300" />
-            <ColumnDefinition Width="*" />
-            <ColumnDefinition Width="400" />
-        </Grid.ColumnDefinitions>
-
-        <Grid.KeyboardAccelerators>
-            <KeyboardAccelerator Key="F" Modifiers="Control" Invoked="FocusSearch_Invoked" />
-            <KeyboardAccelerator Key="M" Modifiers="Menu" Invoked="ToggleNavigation_Invoked" />
-            <KeyboardAccelerator Key="Escape" Invoked="CloseNavigation_Invoked" />
-            <KeyboardAccelerator Key="F5" Invoked="RefreshAccelerator_Invoked" />
-            <KeyboardAccelerator Key="F6" Invoked="ForwardAccelerator_Invoked" />
-            <KeyboardAccelerator Key="F7" Invoked="ForwardAccelerator_Invoked" />
-            <KeyboardAccelerator Key="F8" Invoked="ForwardAccelerator_Invoked" />
-            <KeyboardAccelerator Key="Delete" Invoked="ForwardAccelerator_Invoked" />
-            <KeyboardAccelerator Key="F6" Modifiers="Shift" Invoked="ForwardAccelerator_Invoked" />
-            <KeyboardAccelerator Key="R" Modifiers="Control" Invoked="ForwardAccelerator_Invoked" />
-            <KeyboardAccelerator Key="Q" Modifiers="Control" Invoked="ForwardAccelerator_Invoked" />
-            <KeyboardAccelerator Key="Tab" Invoked="ForwardAccelerator_Invoked" />
-        </Grid.KeyboardAccelerators>
-
-        <views:FiltersNavPane Grid.Row="0"
-                              Grid.Column="0"
-                              HorizontalAlignment="Stretch"
-                              VerticalAlignment="Stretch" />
-
-        <ContentPresenter Grid.Row="0"
-                          Grid.Column="1"
-                          Margin="12"
-                          Content="{Binding CurrentContent}" />
-
-        <Border x:Name="PreviewDetailPane"
-                Grid.Row="0"
-                Grid.Column="2"
-                Margin="12"
-                Padding="12"
-                Background="{ThemeResource CardBackgroundFillColorDefaultBrush}">
-            <ContentPresenter Content="{Binding CurrentDetail}" />
-        </Border>
-
-        <Grid x:Name="StatusBar"
-              Grid.Row="1"
-              Grid.ColumnSpan="3"
-              Background="{ThemeResource LayerOnAcrylicFillColorDefaultBrush}"
-              Padding="8">
-            <winui:InfoBar IsOpen="{Binding IsInfoBarOpen}"
-                           Severity="{Binding HasError, Converter={StaticResource BoolToSeverityConverter}}"
-                           Message="{Binding StatusMessage}" />
-        </Grid>
-
-        <Grid x:Name="NavigationOverlay"
-              Background="{ThemeResource LayerOnAcrylicDefaultBrush}"
-              Visibility="{Binding IsNavOpen, Converter={StaticResource BoolToVisibilityEx}}"
-              IsHitTestVisible="{Binding IsNavOpen}"
-              Tapped="NavigationOverlay_Tapped">
-            <winui:NavigationView x:Name="OverlayNavView"
-                                  Width="320"
-                                  HorizontalAlignment="Left"
-                                  VerticalAlignment="Stretch"
-                                  IsSettingsVisible="True"
-                                  PaneDisplayMode="LeftMinimal"
-                                  IsPaneOpen="{Binding IsNavOpen, Mode=TwoWay}"
-                                  SelectedItem="{Binding SelectedNavItem, Mode=TwoWay}">
-                <winui:NavigationView.MenuItems>
-                    <winui:NavigationViewItem Content="Soubory" Icon="Folder" Tag="Files" />
-                    <winui:NavigationViewItem Content="Import" Icon="Upload" Tag="Import" />
-                    <winui:NavigationViewItem Content="Nastavení" Icon="Setting" Tag="Settings" />
-                </winui:NavigationView.MenuItems>
-            </winui:NavigationView>
-        </Grid>
-    </Grid>
-</UserControl>
